+++ src/features/basic/index.ts
@@ -56,6 +56,7 @@ import './lm-highlight-own-ads';
 import './lm-item-icons/lm-item-icons';
 import './macos-antialiased-font';
 import './mat-linkify-category';
+import './mat-refined-prun-price';
 import './mtra-auto-focus-amount';
 import './mtra-transfer-on-enter';
 import './mu-fix-sector-names';